编码器连续读三次读到对的，自动校正偏置

+++ new-posys/Action_User/baseConfig/icm_20608_g.c
@@ -217,7 +217,7 @@ void icm_update_acc(int gyroNum)
 	}
 	
 }
-void icm_read_accel_acc(float data[GYRO_NUMBER])
+void icm_read_accel_acc(double data[GYRO_NUMBER])
 {
   (data)[0]=acc[0];
   (data)[1]=acc[1];
